chore(tests): move tests to the tests folder
